Implement a more secure channel handshake and fixed a bunch of issues. (#11)
The client now passes the sha-256 hash of the generated ID to the IFrame URL. The provider frame then waits for the ID to be sent via postMessage as the challenge for the handshake. It computes the sha-256 of the nonce challenge received to assert that the sender has the correct ID.

This allows to protect against an attacker that would start the handshake before the client's library.

Fixed issues with demo apps (bad CSS imports, provider URL not settable in settings).
Fixed issue with npm run check that wouldn't compile the custom TSLint rule and hence not be able to use this rule.